Update bcrypt cost factor
Bcrypt has a configurable cost factor to increase the difficulty of brute-force password guessing. It is recommended that this cost is increased over time as computation power increases. A cost of 12 takes around 250ms on a modern computer, which is acceptable in most cases. A minimum cost of 10 is recommended nowadays.